+++ HISTORY.md
@@ -16,6 +16,7 @@
     * new "ide_translation_dict", "ide_set_translation_dict(data)" and "txt_export(txt, name)" extension primitives
 * **Notable Changes:**
     * IF is now variadic and expandable with "else if" cases
+    * explicit variadic input list visuals
     * "this script" reporter has been morphed into the new general "this" introspection/self reflection reporter
     * RUN/CC and CALL/CC primitives have been deprecated and removed from the palette, will still be present and work in existing projects
     * changed the first input slot of the "reshape" block from "list" type to "any" type
@@ -51,6 +52,21 @@
     * German
     * Greek, thanks, HM100!
     * Armenian, thanks to the contributors!
+
+### 2023-06-01
+* blocks: turned on "explicit input lists" for variadic slots
+* blocks: tweaked variadic collapse labels for JOIN and RESHAPE
+
+### 2023-05-31
+* blocks: added collapse labels for explicit input lists
+* blocks: made variadic infix inputs collapsible for explicit input lists
+* blocks: added ability to specify empty variadic collapse labels
+* blocks: tweaked variadic slot arrows layout
+* threads: fixed reportAssociativeBool() for empty variadic inputs
+
+### 2023-05-30
+* blocks: enable explicit input lists in variadic slots, experimental, under construction
+* gui: added hidden experimental "explicit input lists" session setting
 
 ### 2023-05-24
 * threads: refactored variable error handling for new translation mechanism